Organize project structure and update import path accordingly
