+++ src/App.tsx
@@ -1,14 +1,14 @@
 import "./styles/App.css";
 import { BrowserRouter } from "react-router-dom";
-import Home from "./sections/Home";
+import Home from "./sections/Home/Home";
 import NavBar from "./layouts/NavBar";
 import ShootingStarsBackground from "./components/ShootingStarsBackground";
-import TechSkillSection from "./sections/TechSkillSection";
-import ProjectSection from "./sections/ProjectSection";
-import ServiceSection from "./sections/ServiceSection";
-import CredentialSection from "./sections/CredentialSection";
-import TechBlogSection from "./sections/TechBlogSection";
-import ContactSection from "./sections/ContactSection";
+import TechSkillSection from "./sections/TechSkill/TechSkillSection";
+import ProjectSection from "./sections/Project/ProjectSection";
+import ServiceSection from "./sections/Service/ServiceSection";
+import CredentialSection from "./sections/Credential/CredentialSection";
+import TechBlogSection from "./sections/TechBlog/TechBlogSection";
+import ContactSection from "./sections/Contact/ContactSection";
 import Footer from "./layouts/Footer";
 import { useEffect, useState } from "react";
 import LoadingScreen from "./loadingScreen";
