+++ src/sections/Contact/components/Globe.tsx
@@ -1,6 +1,6 @@
 import createGlobe, { COBEOptions } from "cobe";
 import { useCallback, useEffect, useRef } from "react";
-import { cn } from "../lib/utils";
+import { cn } from "../../../lib/utils";
 
 const GLOBE_CONFIG: COBEOptions = {
   width: 800,
