+++ src/sections/Home/components/NumbersCounter.tsx
