+++ src/sections/Home/components/ui/numbers-counter-animation.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useRef } from "react";
 import { useInView, useMotionValue, useSpring } from "framer-motion";
-import { cn } from "../../lib/utils";
+import { cn } from "../../../../lib/utils";
 
 export function NumbersCounterAnimation({
   value,
