+++ src/sections/Project/components/ProjectTabs.tsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
 import { motion } from "framer-motion";
-import { cn } from "../lib/utils";
+import { cn } from "../../../lib/utils";
 
 interface TabProps {
   text: string;
