+++ src/sections/TechBlog/TechBlogSection.tsx
@@ -1,13 +1,12 @@
 import { useMediaQuery } from "react-responsive";
-import ReactHookForm from "../components/ReactHookForm";
+import SwiperBlog from "./components/SwiperBlog";
 import { motion } from "framer-motion";
-import ContactWithGlobe from "../components/ContactWithGlobe";
 
-const ContactSection = () => {
+const TechBlogSection = () => {
   const isMobile = useMediaQuery({ query: "(max-width: 767px)" });
 
   return (
-    <section id="contact" className="">
+    <section id="tech-blog">
       <motion.hr
         className="my-24"
         initial={{ width: "0%", marginLeft: "50%" }}
@@ -28,38 +27,33 @@ const ContactSection = () => {
         viewport={{ once: true }}
       >
         <p className="sm:text-xl my-2 font-bold font-serif text-yellow-500">
-          Episode VI
+          Episode V
         </p>
         <p className="text-yellow-500 star-wars-font text-[1.75rem] sm:text-4xl md:text-5xl lg:text-[3.25rem]">
-          contact
+          tech blog
         </p>
         <p className="mt-4 mb-6 md:mb-10 text-yellow-500 text-sm sm:text-base">
-          Feel free to get in touch for any inquiries
+          Latest posts on the DEV Community
         </p>
       </motion.div>
-      <div className="grid grid-rows sm:grid-cols-2 lg:space-x-10 sm:w-[97%] md:w-[95%] lg:w-[90%]">
-        {/* Framer motion animation for ReactHookForm */}
-        <motion.div
-          className="sm:order-2"
-          initial={{
-            opacity: 0,
-            x: 200,
-          }}
-          whileInView={{
-            opacity: 1,
-            x: 0,
-          }}
-          transition={{ delay: isMobile ? 0.25 : 0.4, duration: 1.5 }}
-          viewport={{ once: true }}
-        >
-          <ReactHookForm />
-        </motion.div>
-        <div className="flex justify-center">
-          <ContactWithGlobe />
-        </div>
-      </div>
+      <motion.div
+        initial={{
+          opacity: 0,
+          y: isMobile ? 0 : -100,
+          x: isMobile ? -200 : 0,
+        }}
+        whileInView={{
+          opacity: 1,
+          y: 0,
+          x: 0,
+        }}
+        transition={{ delay: isMobile ? 0.25 : 0.4, duration: 1.5 }}
+        viewport={{ once: true }}
+      >
+        <SwiperBlog />
+      </motion.div>
     </section>
   );
 };
 
-export default ContactSection;
+export default TechBlogSection;
